feat: Match 도메인 객체 및 MatchCreator 생성 완료, refactor: imageCount Long->Integer 수정

+++ src/main/java/com/team/moodit/storage/db/core/MatchEntity.java
@@ -30,7 +30,7 @@ public class MatchEntity extends BaseIdEntity {
 
 
     @Column(name = "initial_image_count", nullable = false)
-    private Long initialImageCount;
+    private Integer initialImageCount;
 
 
     @CreationTimestamp
